+++ simmain.cpp
@@ -63,7 +63,7 @@ void SimMain::on_setRAMButton_clicked()
   bool addrOk = false;
   bool valOk = false;
   const uint16_t addr = ui->addrBox->text().toUInt(&addrOk, 0);
-  const uint8_t val = ui->valBox->text().toUInt(&valOk, 0);
+  const uint16_t val = ui->valBox->text().toUInt(&valOk, 0);
   if (addrOk && valOk)
   {
     m_sim.setRAMLoc(addr, val);
